added support for order heap

+++ src/main/java/com/weekday/orderingsystem/ov1/dto/CookingSlot.java
@@ -9,6 +9,6 @@ import lombok.NoArgsConstructor;
 @Data
 public class CookingSlot {
     int orderId;
-    int orderCompletionTime;
+    double orderCompletionTime;
     int slotsUsed;   
 }
